fixed CGs
